Restored command line stuff
